Further clarify/simplify event/interrupt map usage.
This furthers the changes started in fbb1fb0027b827ecbbb34dca18270b7cac8fcbb1.
prussdrv_pru_clear_event and prussdrv_pru_reset_interrupt now explicitly only
clear the event and reset the host interrupt line.  As indicated in
8bfab8826e7800721cf7f6aecdb779a47a17ded2, the order of these operations is
critical so-as not to falsely re-trigger the host interrupt (see See Section
6.4.9 of Reference manual).

This change also adds a function **prussdrv_pru_reset_event**.  This function
first clears the event specified by the user and then resets the host interrupt
line to which the specified event is tied (no need for the user to also specify
the host interrupt line).