fix and make for loops uniform in intc_init
Fix is for when _all_ events, or _all_ channels, or _all_ hosts are used.  Prior
to this fix, the result would have been undefined (dependent on random values in
RAM).

Changed internal format of intc_data to all unsigned.  List terminations now
test for anything that is >= NUM_PRU_{HOSTS, SYS_EVTS, CHANNELS}.